pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "cx-shell"
version = "0.1.0"
description = "The Contextual Shell: a declarative, multi-stage automation platform for modern data and operations teams."
authors = [
    { name = "Syncropel", email = "dev@syncropel.com" },
]
readme = "README.md" # We will need to create this file later
requires-python = ">=3.12"
license = { text = "MIT" }

# --- Core Dependencies (Merged from both projects) ---
dependencies = [
    # Core CLI and Logging
    "cx-core-schemas @ git+https://github.com/syncropel/cx-core-schemas.git",
    "typer[all]>=0.9.0",
    "structlog>=24.1.0",
    "rich>=13.0.0",
    
    # Core Utilities
    "pydantic>=2.0,<3.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    
    # Connector-specific
    "httpx[http2]>=0.25.0",
    "tiktoken>=0.7.0",
    
    # Transformer-specific
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pyarrow>=15.0.0",
    "jmespath>=1.0.0",
    "networkx>=3.0",
    "prompt-toolkit>=3.0.0",
    "lark>=1.1.9",
    "jsonpatch>=1.33",
    "instructor[openai,anthropic,google-vertexai]>=1.0",
    "tenacity>=8.2.3",
    "lancedb>=0.6.1",
    "fastembed>=0.2.7",
    "pyarrow>=15.0.0", 
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "playwright>=1.42.0",
    "beautifulsoup4>=4.9.0,<5.0.0",
    "markdown>=3.6",
]

# --- Optional Dependencies (Merged) ---
[project.optional-dependencies]
sql = [
    "sqlalchemy[asyncio]>=2.0.0",
    "aioodbc",
    "trino[sqlalchemy]>=0.334.0",
]
git = [
    "GitPython>=3.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-depends>=1.0.1",
    "ruff",
    "pyinstaller>=6.0.0",
]

integrated = [
    "surrealdb==1.0.4",
    "hvac>=2.3.0",
]

all = [
    "cx-shell[sql,git,dev,integrated]"
]

# --- This is the new, unified CLI entry point ---
[project.scripts]
cx = "cx_shell.main:main"

# --- Package Discovery ---
[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
    "network: marks tests that require a live network connection",
]